tweedle/api/requirements.txt
asgiref==3.4.1
Deprecated==1.2.13
Django==4.0.1
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
packaging==21.3
psycopg2==2.8.6
PyJWT==2.3.0
pyparsing==3.0.7
pytz==2022.1
redis==4.1.0
sqlparse==0.4.2
tzdata==2021.5
wrapt==1.14.0
celery==5.2.6
gunicorn==20.1.0
firebase-admin==5.2.0